// src/app/core/api/model/createReservationRequestDto.ts
/**
 * AGORA API
 *
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
/**
 * Cote serveur : LocalTime JSON (ex. "09:00:00").
 * Utiliser des chaines pour eviter la serialisation { hour, minute } du client OpenAPI.
 */
export interface CreateReservationRequestDto {
  resourceId: string;
  date: string;
  slotStart: string;
  slotEnd: string;
  purpose: string;
  groupId?: string;
}